% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_boost.R
\name{control_boost}
\alias{control_boost}
\title{Auxiliary Function for Controlling the Gradient-Boosting Estimation}
\usage{
control_boost(
  maxit = 6000,
  nu = 0.1,
  mstop = "aic",
  nfolds = 10,
  foldid = NULL,
  cores = 1L
)
}
\arguments{
\item{maxit}{integer; giving the number of initial boosting iterations. Default: \code{maxit = 6000}.}

\item{nu}{double; defining the step size or shrinkage parameter. Default: \code{nu = 0.1}.}

\item{mstop}{character; stopping criterion "\code{max}", "\code{aic}" (default), "\code{bic}", "\code{cv}".}

\item{nfolds}{integer; number of folds in cross validation.}

\item{foldid}{vector; an optional vector of values between 1 and \code{nfolds} identifying the fold for each observation.}

\item{cores}{integer; the number of cores used for cross validation. Default: \code{cores = 1L}.}
}
\value{
An object of class \code{mixnhreg.boost}.
}
\description{
Auxiliary function for controlling the non-cylcic gradient-boosting estimation (see references).
}
