% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_optim.R
\name{control_optim}
\alias{control_optim}
\title{Auxiliary Function for Controlling Optimization}
\usage{
control_optim(
  method = "BFGS",
  maxit = 100,
  start = list(location = NULL, scale = NULL, weight = NULL),
  hessian = FALSE,
  trace = FALSE,
  lower = -Inf,
  upper = Inf
)
}
\arguments{
\item{method}{character; optimization method passed to \code{\link{optim}}. Default: "\code{BFGS}".}

\item{maxit}{integer; maximum number of iterations. Default: \code{100}.}

\item{start}{list; list of initial values for the location, scale and weight parameters.
If a parameter is set to \code{NULL} (default), corresponding initial values are internally determined.}

\item{hessian}{logical; if \code{TRUE} the numerical Hessian matrix from the \code{\link{optim}} output is used for the covariance matrix estimation.
If \code{FALSE} (default) no covariance matrix is determined.}

\item{trace}{logical; if \code{TRUE}, the optimization is traced, otherwise \code{FALSE} (default) not.}

\item{lower}{integer; lower bound for the "\code{L-BFGS-B}", "\code{Brent}" methods.}

\item{upper}{integer; upper bound for the "\code{L-BFGS-B}", "\code{Brent}" methods.}
}
\value{
An object of class \code{mixnhreg.optim}.
}
\description{
Auxiliary function for controlling the optimization.
}
