% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixnorm.R
\name{Mixture_Normal}
\alias{Mixture_Normal}
\alias{dmixnorm}
\alias{pmixnorm}
\alias{qmixnorm}
\alias{rmixnorm}
\title{The Mixture Normal Distribution}
\usage{
dmixnorm(x, location, scale, weight, log = FALSE)

pmixnorm(q, location, scale, weight, lower.tail = TRUE, log.p = FALSE)

qmixnorm(p, location, scale, weight, lower.tail = TRUE, log.p = FALSE)

rmixnorm(n, location, scale, weight)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{location}{matrix of location parameters.}

\item{scale}{matrix of scale parameters.}

\item{weight}{matrix of weight parameters.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities p are given as log(p).}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation for the mixture normal distribution
as well as the corresponding \code{family.mixnhreg} object.
}
