% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mixnhreg.boost.R
\name{plot.mixnhreg.boost}
\alias{plot.mixnhreg.boost}
\title{Plot Standardized Coefficient and log-likelihood Paths}
\usage{
\method{plot}{mixnhreg.boost}(x, type = "location", loglik = FALSE, mstop = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{mixnhreg.boost}.}

\item{type}{character; specifies for which distribution parameters, i.e. "\code{location}" (default), "\code{scale}", "\code{weight}" the paths should be created.}

\item{loglik}{logical; if \code{TRUE}, the log-likelihood contribution path is plotted instead of the standardized coefficient path. Default: \code{FALSE}.}

\item{mstop}{logical; if \code{TRUE} (default) a vertical line for the optimal number of boosting iterations is plotted, otherwise \code{FALSE} not.}

\item{...}{unused.}
}
\description{
This function plots standardized coefficient and log-likelihood paths of an \code{mixnhreg.boost} object.
}
\examples{
# load data
data("station")

# fit mixture normal distribution with two components via gradient-boosting
(fit_boost <- mixnhreg(formula = obs ~ sin1 + cos1 + temp_mean | temp_ctrl,
                       scale.formula = ~ sin1 + cos1 + log(temp_sd) | 1,
                       weight.formula = ~ sin1 + cos1 | 1,
                       data = station,
                       control = control_boost(mstop = "cv")))

# standardized coefficient path for location parameter
plot(fit_boost, type = "location")

# log-likelihood contribution path for location parameter
plot(fit_boost, type = "location", loglik = TRUE)

}
