% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_mixnorm.R
\name{scores_mixnorm}
\alias{scores_mixnorm}
\alias{logs_mixnorm}
\alias{crps_mixnorm}
\title{Scoring Rules for Mixture Normal Distribution}
\usage{
logs_mixnorm(x, location, scale, weight)

crps_mixnorm(x, location, scale, weight)
}
\arguments{
\item{x}{vector of quantiles.}

\item{location}{matrix of location parameters.}

\item{scale}{matrix of scale parameters.}

\item{weight}{matrix of weight parameters.}
}
\description{
Logarithmic score (logs) a.k.a. negative log-likelihood and continuous ranked probability score (crps) for the mixture normal distribution.
}
\references{
Grimit E.P., Gneiting T., Berrocal V., Johnson N.A. (2006). The continuous ranked probability score for circular variables and its application to mesoscale forecast ensemble verification.
Quarterly Journal of the Royal Meteorological Society, 132, pp. 2925–2942. doi: <https://doi.org/10.1256/qj.05.235>.
}
