% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mstop.R
\name{update_mstop}
\alias{update_mstop}
\title{Auxiliary Function for Updating the Stopping Criterion in the Gradient-Boosting Estimation}
\usage{
update_mstop(object, mstop = "aic")
}
\arguments{
\item{object}{an object of class \code{mixnhreg.boost}.}

\item{mstop}{character; new stopping criterion "\code{max}", "\code{aic}" (default), "\code{bic}", "\code{cv}".}
}
\value{
An object of class \code{mixnhreg}.
}
\description{
Auxiliary function for updating the stopping criterion "\code{mstop}" and consequently an \code{mixnhreg.boost} object
based on the initial settings provided for the "\code{control}" argument in \code{\link{mixnhreg}}.
}
\examples{
# load data
data("station")

# fit mixture normal distribution with two components via gradient-boosting
(fit_boost1 <- mixnhreg(formula = obs ~ sin1 + cos1 + temp_mean | temp_ctrl,
                        scale.formula = ~ sin1 + cos1 + log(temp_sd) | 1,
                        weight.formula = ~ sin1 + cos1 | 1,
                        data = station,
                        control = control_boost(mstop = "cv")))

# update stopping criterion to "bic"
(fit_boost2 <- update_mstop(fit_boost1, mstop = "bic"))

}
